+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add Samsung browser.
+
 ## 4.0.0
 
 - Add Chrome Lighthouse to bot list.
@@ -8,7 +12,6 @@
   will be returned for versions prior to 10.12, and "macOS" for newer versions.
 - Remove `Browser.modern_rules` and `Browser::Base#modern?`.
 - Add DuckDuckGo browser.
-- Add Samsung browser.
 
 # 3.0.3